improve dummy data and tests
- improve the dummy data to be generated when building the mock db
- restructure dummy data folder structure
- remove all real data from the dummy data
- update the tests to use the new dummay data, and also
  - fix the incorrect 404 test
  - add a test for viewing images
