+++ frontend/src/components/MediaGallery.tsx/index.tsx
@@ -34,7 +34,10 @@ export const MediaGallery = component$<Props>(({ medias }) => {
 	return (
 		<>
 			{!!medias.length && (
-				<div class={`media-gallery overflow-hidden grid gap-1 h-52 md:h-60 lg:h-72 xl:h-80`}>
+				<div
+					data-testid="media-gallery"
+					class={`media-gallery overflow-hidden grid gap-1 h-52 md:h-60 lg:h-72 xl:h-80`}
+				>
 					{medias.map((media) => (
 						<div class="w-full flex items-center justify-center overflow-hidden bg-black">
 							{media.type === 'image' && <Image mediaAttachment={media} onOpenImagesModal$={onOpenImagesModal} />}
