--- frontend/src/dummyData.tsx
@@ -1,1764 +1,0 @@
-import { MastodonLink, MastodonStatus, TagDetails } from './types'
-
-// Raw statuses taken directly from mastodon
-const mastodonRawStatuses: MastodonStatus[] = [
-	{
-		id: '109632280734155646',
-		created_at: '2023-01-04T17:53:39.000Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://universeodon.com/users/georgetakei/statuses/109632213422003756',
-		url: 'https://universeodon.com/@georgetakei/109632213422003756',
-		replies_count: 155,
-		reblogs_count: 705,
-		favourites_count: 329,
-		edited_at: null,
-		content: "\u003cp\u003eFine. I'll be House Speaker.\u003c/p\u003e",
-		reblog: null,
-		account: {
-			id: '109355700962815786',
-			username: 'georgetakei',
-			acct: 'georgetakei@universeodon.com',
-			display_name: 'George Takei :verified: 🏳️‍🌈🖖🏽',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-11-15T00:00:00.000Z',
-			note: '\u003cp\u003eI boldly went to this new site. Follow for more recipes and tips.\u003c/p\u003e',
-			url: 'https://universeodon.com/@georgetakei',
-			avatar:
-				'https://files.mastodon.social/cache/accounts/avatars/109/355/700/962/815/786/original/7d278db7224de27d.jpg',
-			avatar_static:
-				'https://files.mastodon.social/cache/accounts/avatars/109/355/700/962/815/786/original/7d278db7224de27d.jpg',
-			header:
-				'https://files.mastodon.social/cache/accounts/headers/109/355/700/962/815/786/original/01c05d0b46e15480.jpg',
-			header_static:
-				'https://files.mastodon.social/cache/accounts/headers/109/355/700/962/815/786/original/01c05d0b46e15480.jpg',
-			followers_count: 331437,
-			following_count: 37,
-			statuses_count: 187,
-			last_status_at: '2023-01-04',
-			emojis: [
-				{
-					shortcode: 'verified',
-					url: 'https://files.mastodon.social/cache/custom_emojis/images/000/452/462/original/947cae7ac4dfdfa0.png',
-					static_url:
-						'https://files.mastodon.social/cache/custom_emojis/images/000/452/462/static/947cae7ac4dfdfa0.png',
-					visible_in_picker: true,
-				},
-			],
-			fields: [
-				{
-					name: 'Instagram',
-					value:
-						'\u003ca href="https://www.instagram.com/georgehtakei/" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://www.\u003c/span\u003e\u003cspan class=""\u003einstagram.com/georgehtakei/\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Twitter',
-					value:
-						'\u003ca href="https://twitter.com/GeorgeTakei" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003etwitter.com/GeorgeTakei\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Facebook',
-					value:
-						'\u003ca href="https://www.facebook.com/georgehtakei/" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://www.\u003c/span\u003e\u003cspan class=""\u003efacebook.com/georgehtakei/\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109630407170172986',
-		created_at: '2023-01-04T10:14:17.855Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/BethanyBlack/statuses/109630407170172986',
-		url: 'https://mastodon.social/@BethanyBlack/109630407170172986',
-		replies_count: 52,
-		reblogs_count: 630,
-		favourites_count: 1434,
-		edited_at: '2023-01-04T10:15:59.795Z',
-		content: '\u003cp\u003eWe did it!  *wipes tear from eye*\u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Web', website: null },
-		account: {
-			id: '236033',
-			username: 'BethanyBlack',
-			acct: 'BethanyBlack',
-			display_name: 'Bethany Black',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2017-11-13T00:00:00.000Z',
-			note: '\u003cp\u003e‘Comedian’ managed by United Agents. \u003ca href="https://mastodon.social/tags/ActuallyAutistic" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eActuallyAutistic\u003c/span\u003e\u003c/a\u003e and incredibly thin skinned.\u003c/p\u003e\u003cp\u003eProfile pic alt text: photo of me, white butch lesbian with a quiff, and a cheeky smile. wearing a burgundy hoodie and a grey tshirt\u003c/p\u003e\u003cp\u003eOn Tour in 2023 with my show “Always Bet On Beth” check webshite for details\u003c/p\u003e',
-			url: 'https://mastodon.social/@BethanyBlack',
-			avatar: 'https://files.mastodon.social/accounts/avatars/000/236/033/original/39a4260e678b3dfb.png',
-			avatar_static: 'https://files.mastodon.social/accounts/avatars/000/236/033/original/39a4260e678b3dfb.png',
-			header: 'https://files.mastodon.social/accounts/headers/000/236/033/original/1d93c5afb5939257.jpeg',
-			header_static: 'https://files.mastodon.social/accounts/headers/000/236/033/original/1d93c5afb5939257.jpeg',
-			followers_count: 2344,
-			following_count: 205,
-			statuses_count: 2027,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [
-				{
-					name: 'Website',
-					value:
-						'\u003ca href="https://www.bethanyblack.co.uk" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://www.\u003c/span\u003e\u003cspan class=""\u003ebethanyblack.co.uk\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2022-12-22T10:01:27.713+00:00',
-				},
-			],
-		},
-		media_attachments: [
-			{
-				id: '109630397257532036',
-				type: 'image',
-				url: 'https://files.mastodon.social/media_attachments/files/109/630/397/257/532/036/original/5324673ca7d402c7.jpeg',
-				preview_url:
-					'https://files.mastodon.social/media_attachments/files/109/630/397/257/532/036/small/5324673ca7d402c7.jpeg',
-				remote_url: null,
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					focus: { x: -0.02, y: 0.87 },
-					original: { width: 1158, height: 1790, size: '1158x1790', aspect: 0.646927374301676 },
-					small: { width: 386, height: 597, size: '386x597', aspect: 0.6465661641541038 },
-				},
-				description:
-					'A PC Gamer article with the headline:Good job, internet: You bullied NFTs out of mainstream games.\nOr at the very least, made fun of them enough to keep us all from losing our grip on reality.\n\nThe picture accompanying the article is Seth Green’s Ape NFT',
-				blurhash: 'UaMQt_aetSj].8j[WAay4noLjtj[00j[Wray',
-			},
-		],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109632119277219029',
-		created_at: '2023-01-04T17:29:42.531Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/twostraws/statuses/109632119277219029',
-		url: 'https://mastodon.social/@twostraws/109632119277219029',
-		replies_count: 11,
-		reblogs_count: 204,
-		favourites_count: 297,
-		edited_at: null,
-		content: '\u003cp\u003eHappy April Fools day to everyone who thinks the US date format is weird 🎉\u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Mastodon Twitter Crossposter', website: null },
-		account: {
-			id: '413899',
-			username: 'twostraws',
-			acct: 'twostraws',
-			display_name: 'Paul Hudson',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2018-08-11T00:00:00.000Z',
-			note: '\u003cp\u003eI write Swift and write about Swift, but most people prefer my dogs.\u003c/p\u003e',
-			url: 'https://mastodon.social/@twostraws',
-			avatar: 'https://files.mastodon.social/accounts/avatars/000/413/899/original/b4b55495ddda94a9.jpeg',
-			avatar_static: 'https://files.mastodon.social/accounts/avatars/000/413/899/original/b4b55495ddda94a9.jpeg',
-			header: 'https://static-cdn.mastodon.social/headers/original/missing.png',
-			header_static: 'https://static-cdn.mastodon.social/headers/original/missing.png',
-			followers_count: 4600,
-			following_count: 59,
-			statuses_count: 106,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [],
-		},
-		media_attachments: [],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109631525794963716',
-		created_at: '2023-01-04T14:58:14.000Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://infosec.exchange/users/dnsprincess/statuses/109631523691245767',
-		url: 'https://infosec.exchange/@dnsprincess/109631523691245767',
-		replies_count: 85,
-		reblogs_count: 380,
-		favourites_count: 115,
-		edited_at: null,
-		content: '\u003cp\u003ethis is a relaxing post\u003c/p\u003e',
-		reblog: null,
-		account: {
-			id: '109274277494505675',
-			username: 'dnsprincess',
-			acct: 'dnsprincess@infosec.exchange',
-			display_name: 'alissa',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-11-02T00:00:00.000Z',
-			note: '\u003cp\u003ebe good, don\'t die\u003cbr\u003e\u003ca href="https://infosec.exchange/tags/Wardriving" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003eWardriving\u003c/span\u003e\u003c/a\u003e | Hacking | Shenanigans\u003cbr\u003eAsk me about \u003cspan class="h-card"\u003e\u003ca href="https://infosec.exchange/@CircleCityCon" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003eCircleCityCon\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e\u003cbr\u003enon-binary, they/them 🏳️‍🌈\u003c/p\u003e',
-			url: 'https://infosec.exchange/@dnsprincess',
-			avatar:
-				'https://files.mastodon.social/cache/accounts/avatars/109/274/277/494/505/675/original/82e61596d5430854.jpg',
-			avatar_static:
-				'https://files.mastodon.social/cache/accounts/avatars/109/274/277/494/505/675/original/82e61596d5430854.jpg',
-			header:
-				'https://files.mastodon.social/cache/accounts/headers/109/274/277/494/505/675/original/acd0335233b97e5a.jpg',
-			header_static:
-				'https://files.mastodon.social/cache/accounts/headers/109/274/277/494/505/675/original/acd0335233b97e5a.jpg',
-			followers_count: 4730,
-			following_count: 502,
-			statuses_count: 1228,
-			last_status_at: '2023-01-04',
-			emojis: [],
-			fields: [
-				{
-					name: 'My bio',
-					value:
-						'\u003ca href="https://bio.fm/alissa" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003ebio.fm/alissa\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Blog',
-					value:
-						'\u003ca href="https://dnsprincess.com" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003ednsprincess.com\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Twitter',
-					value:
-						'\u003ca href="https://twitter.com/dnsprincess" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003etwitter.com/dnsprincess\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [
-			{
-				id: '109631525757025938',
-				type: 'image',
-				url: 'https://files.mastodon.social/cache/media_attachments/files/109/631/525/757/025/938/original/81d3d9d14ebae6e5.jpg',
-				preview_url:
-					'https://files.mastodon.social/cache/media_attachments/files/109/631/525/757/025/938/small/81d3d9d14ebae6e5.jpg',
-				remote_url:
-					'https://media.infosec.exchange/infosecmedia/media_attachments/files/109/631/507/229/477/156/original/91f559a9cebf063d.jpg',
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					focus: { x: 0.0, y: 0.0 },
-					original: { width: 1000, height: 667, size: '1000x667', aspect: 1.4992503748125936 },
-					small: { width: 588, height: 392, size: '588x392', aspect: 1.5 },
-				},
-				description:
-					"A gorgeous display of cable management. The cable intake is on the upper left and the server is to the right. The cables drop in, in organized bundles. Blue, green, yellow, and orange. The neat cables and tied together with their respective color. The green color wraps near the yellow, but sinks to the floor. The blue is the largest section, and makes a cute lil loop around the orange cables before seating into the top three servers. The orange is in the middle surrounding the yellow, and the yellow is the majority of the middle section. \nThe organized cables bounded together almost look like a water park slide. It's organized, beautiful, and satisfying. \nThe cables appreciate you.",
-				blurhash: 'UFC~|#.79ZX4.T%LITRkbyjIIAWU~9WB9GNH',
-			},
-		],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109632154173927159',
-		created_at: '2023-01-04T17:29:16.000Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://tilde.zone/users/fivetonsflax/statuses/109632117552890130',
-		url: 'https://tilde.zone/@fivetonsflax/109632117552890130',
-		replies_count: 24,
-		reblogs_count: 241,
-		favourites_count: 38,
-		edited_at: null,
-		content:
-			'\u003cp\u003eWhat fresh hell is this? Adobe has opted me in to letting them train their algorithms on my photos? That’s the end of me using their cloud storage for anything, ever. \u003ca href="https://tilde.zone/tags/privacy" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003eprivacy\u003c/span\u003e\u003c/a\u003e \u003ca href="https://tilde.zone/tags/adobe" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003eadobe\u003c/span\u003e\u003c/a\u003e \u003ca href="https://tilde.zone/tags/lightroom" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003elightroom\u003c/span\u003e\u003c/a\u003e \u003ca href="https://tilde.zone/tags/LightroomMobile" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003eLightroomMobile\u003c/span\u003e\u003c/a\u003e\u003c/p\u003e',
-		reblog: null,
-		account: {
-			id: '109288055954478061',
-			username: 'fivetonsflax',
-			acct: 'fivetonsflax@tilde.zone',
-			display_name: 'Ben Rosengart',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-11-04T00:00:00.000Z',
-			note: '\u003cp\u003eAmbivalent booster\u003c/p\u003e',
-			url: 'https://tilde.zone/@fivetonsflax',
-			avatar:
-				'https://files.mastodon.social/cache/accounts/avatars/109/288/055/954/478/061/original/d9c243a7a8d8ef6a.png',
-			avatar_static:
-				'https://files.mastodon.social/cache/accounts/avatars/109/288/055/954/478/061/original/d9c243a7a8d8ef6a.png',
-			header: 'https://static-cdn.mastodon.social/headers/original/missing.png',
-			header_static: 'https://static-cdn.mastodon.social/headers/original/missing.png',
-			followers_count: 124,
-			following_count: 187,
-			statuses_count: 621,
-			last_status_at: '2023-01-04',
-			emojis: [],
-			fields: [],
-		},
-		media_attachments: [
-			{
-				id: '109632154049976745',
-				type: 'image',
-				url: 'https://files.mastodon.social/cache/media_attachments/files/109/632/154/049/976/745/original/7539741f0d0a96d7.jpeg',
-				preview_url:
-					'https://files.mastodon.social/cache/media_attachments/files/109/632/154/049/976/745/small/7539741f0d0a96d7.jpeg',
-				remote_url:
-					'https://media.tilde.zone/system/media_attachments/files/109/632/113/897/812/481/original/5d4a34483c33ea81.jpeg',
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					original: { width: 1668, height: 861, size: '1668x861', aspect: 1.9372822299651569 },
-					small: { width: 668, height: 345, size: '668x345', aspect: 1.936231884057971 },
-				},
-				description:
-					'Screenshot of adobe.com showing that “Content analysis” is turned on, and summarizing how Adobe is invading my privacy without my consent.',
-				blurhash: 'UAR:HGjE%Ms:Ris:ayoe~qV@Iot7?bV@Rjof',
-			},
-		],
-		mentions: [],
-		tags: [
-			{ name: 'privacy', url: 'https://mastodon.social/tags/privacy' },
-			{ name: 'adobe', url: 'https://mastodon.social/tags/adobe' },
-			{ name: 'lightroom', url: 'https://mastodon.social/tags/lightroom' },
-			{ name: 'lightroommobile', url: 'https://mastodon.social/tags/lightroommobile' },
-		],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109632224523439398',
-		created_at: '2023-01-04T17:42:09.000Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://social.growyourown.services/users/FediFollows/statuses/109632168225800814',
-		url: 'https://social.growyourown.services/@FediFollows/109632168225800814',
-		replies_count: 18,
-		reblogs_count: 239,
-		favourites_count: 26,
-		edited_at: null,
-		content:
-			'\u003cp\u003eAs you might have read today, \u003ca href="https://social.growyourown.services/tags/Adobe" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003eAdobe\u003c/span\u003e\u003c/a\u003e is doing really bad stuff (\u003ca href="https://toot.cafe/@baldur/109630505660962387" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class="ellipsis"\u003etoot.cafe/@baldur/109630505660\u003c/span\u003e\u003cspan class="invisible"\u003e962387\u003c/span\u003e\u003c/a\u003e).\u003c/p\u003e\u003cp\u003eFortunately there are many projects you can follow on here which are providing free open source alternatives:\u003c/p\u003e\u003cp\u003e➡️ \u003cspan class="h-card"\u003e\u003ca href="https://mastodon.social/@darktable" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003edarktable\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e - Photo software which lets you organise, view and develop raw image files\u003c/p\u003e\u003cp\u003e➡️ \u003cspan class="h-card"\u003e\u003ca href="https://mastodon.art/@Krita" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003eKrita\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e  - Professional digital painting \u0026amp; 2D animation app\u003c/p\u003e\u003cp\u003e➡️ \u003cspan class="h-card"\u003e\u003ca href="https://floss.social/@kdenlive" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003ekdenlive\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e - NLE video editing software\u003c/p\u003e\u003cp\u003e➡️ \u003cspan class="h-card"\u003e\u003ca href="https://mastodon.art/@inkscape" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003einkscape\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e - Vector illustration software\u003c/p\u003e\u003cp\u003e➡️ \u003cspan class="h-card"\u003e\u003ca href="https://mastodon.social/@Blender" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003eBlender@mastodon.social\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e \u0026amp; \u003cspan class="h-card"\u003e\u003ca href="https://video.blender.org/accounts/blender" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003eblender@video.blender.org\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e - 3D animation and special effects\u003c/p\u003e',
-		reblog: null,
-		account: {
-			id: '109390969127442952',
-			username: 'FediFollows',
-			acct: 'FediFollows@social.growyourown.services',
-			display_name: 'FediFollows ☃️',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-11-22T00:00:00.000Z',
-			note: '\u003cp\u003eHighlighting interesting accounts to follow on Mastodon and the Fediverse 🌐\u003c/p\u003e\u003cp\u003eFor recommendations organised by topic, see my website at \u003ca href="https://fedi.directory" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003efedi.directory\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e\u003c/p\u003e\u003cp\u003eI am a volunteer, you can buy me a coffee at \u003ca href="https://ko-fi.com/fedithing" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003eko-fi.com/fedithing\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e or become a patron at \u003ca href="https://liberapay.com/FediThing" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003eliberapay.com/FediThing\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e\u003c/p\u003e\u003cp\u003eMaintained by \u003cspan class="h-card"\u003e\u003ca href="https://tech.lgbt/@FediThing" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003eFediThing\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e\u003c/p\u003e\u003cp\u003e🎂 Account originally joined mastodon.online on 11th June 2020\u003c/p\u003e',
-			url: 'https://social.growyourown.services/@FediFollows',
-			avatar:
-				'https://files.mastodon.social/cache/accounts/avatars/109/390/969/127/442/952/original/c6989e36027f5ef7.png',
-			avatar_static:
-				'https://files.mastodon.social/cache/accounts/avatars/109/390/969/127/442/952/original/c6989e36027f5ef7.png',
-			header:
-				'https://files.mastodon.social/cache/accounts/headers/109/390/969/127/442/952/original/cde315aee1c9305b.png',
-			header_static:
-				'https://files.mastodon.social/cache/accounts/headers/109/390/969/127/442/952/original/cde315aee1c9305b.png',
-			followers_count: 47280,
-			following_count: 2,
-			statuses_count: 67,
-			last_status_at: '2023-01-04',
-			emojis: [],
-			fields: [
-				{
-					name: 'Website',
-					value:
-						'\u003ca href="https://fedi.directory" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003efedi.directory\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2023-01-04T05:59:30.169+00:00',
-				},
-				{
-					name: 'See also',
-					value:
-						'\u003cspan class="h-card"\u003e\u003ca href="https://mstdn.social/@feditips" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003efeditips@mstdn.social\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e, \u003cspan class="h-card"\u003e\u003ca href="https://social.growyourown.services/@FediVideo" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003eFediVideo\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e, \u003cspan class="h-card"\u003e\u003ca href="https://social.growyourown.services/@homegrown" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003ehomegrown\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e \u0026amp; \u003cspan class="h-card"\u003e\u003ca href="https://tech.lgbt/@FediThing" class="u-url mention" rel="nofollow noopener noreferrer" target="_blank"\u003e@\u003cspan\u003eFediThing@tech.lgbt\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Buy me a coffee ☕',
-					value:
-						'\u003ca href="https://ko-fi.com/fedithing" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003eko-fi.com/fedithing\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Become a patron 🔁',
-					value:
-						'\u003ca href="https://liberapay.com/FediThing" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003eliberapay.com/FediThing\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [],
-		mentions: [
-			{ id: '1303783', username: 'darktable', url: 'https://mastodon.social/@darktable', acct: 'darktable' },
-			{ id: '298954', username: 'krita', url: 'https://mastodon.art/@Krita', acct: 'krita@mastodon.art' },
-			{
-				id: '109137203193899478',
-				username: 'kdenlive',
-				url: 'https://floss.social/@kdenlive',
-				acct: 'kdenlive@floss.social',
-			},
-			{ id: '318492', username: 'inkscape', url: 'https://mastodon.art/@inkscape', acct: 'inkscape@mastodon.art' },
-			{ id: '284692', username: 'Blender', url: 'https://mastodon.social/@Blender', acct: 'Blender' },
-			{
-				id: '376434',
-				username: 'blender',
-				url: 'https://video.blender.org/accounts/blender',
-				acct: 'blender@video.blender.org',
-			},
-		],
-		tags: [{ name: 'adobe', url: 'https://mastodon.social/tags/adobe' }],
-		emojis: [],
-		card: {
-			url: 'https://toot.cafe/@baldur/109630505660962387',
-			title: 'Baldur Bjarnason (@baldur@toot.cafe)',
-			description:
-				"Attached: 1 image\n\nTurns out that Adobe is collecting all of its customers' pictures into a machine learning training set.\n\nThis is opt-out, not opt-in so if you use Lightroom, for example, it defaults to adding all of your photos to the set.\n\nIf these are unpublished pictures, work-in-progress, etc. they'll still be analysed as soon as they're synced.\n\nI've been using Lightroom to sync photos from my Windows desktop to my iPad. Now I need to reconsider that.",
-			type: 'link',
-			author_name: '',
-			author_url: '',
-			provider_name: 'Toot Café',
-			provider_url: '',
-			html: '',
-			width: 400,
-			height: 80,
-			image: 'https://files.mastodon.social/cache/preview_cards/images/051/362/766/original/3ad77217a326f17e.png',
-			embed_url: '',
-			blurhash: 'UBR{#?M{Rjof4nR*WBRjRioLj@WB~qaeWCof',
-		},
-		poll: null,
-	},
-	{
-		id: '109632174659610537',
-		created_at: '2023-01-04T17:43:47.602Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/scottjshapiro/statuses/109632174659610537',
-		url: 'https://mastodon.social/@scottjshapiro/109632174659610537',
-		replies_count: 16,
-		reblogs_count: 71,
-		favourites_count: 173,
-		edited_at: null,
-		content: '\u003cp\u003eBREAKING: Black smoke coming from Capitol chimney.\u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Mastodon for iOS', website: 'https://app.joinmastodon.org/ios' },
-		account: {
-			id: '109243342914018772',
-			username: 'scottjshapiro',
-			acct: 'scottjshapiro',
-			display_name: 'Scott Shapiro',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-10-28T00:00:00.000Z',
-			note: '\u003cp\u003eSouthmayd Prof @YaleLawSch + Philosophy \u003cspan class="h-card"\u003e\u003ca href="https://mastodon.social/@yale" class="u-url mention"\u003e@\u003cspan\u003eyale\u003c/span\u003e\u003c/a\u003e\u003c/span\u003e. Ed, @LegalTheory + Stanford Encyclopedia of Phil. \u0026quot;Fancy Bear Goes Phishing\u0026quot;—forthcoming, May 23, 2023\u003c/p\u003e',
-			url: 'https://mastodon.social/@scottjshapiro',
-			avatar: 'https://files.mastodon.social/accounts/avatars/109/243/342/914/018/772/original/f64c5d876eea3e1d.jpeg',
-			avatar_static:
-				'https://files.mastodon.social/accounts/avatars/109/243/342/914/018/772/original/f64c5d876eea3e1d.jpeg',
-			header: 'https://files.mastodon.social/accounts/headers/109/243/342/914/018/772/original/f13e26fc650eddd2.jpeg',
-			header_static:
-				'https://files.mastodon.social/accounts/headers/109/243/342/914/018/772/original/f13e26fc650eddd2.jpeg',
-			followers_count: 26403,
-			following_count: 87,
-			statuses_count: 400,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [],
-		},
-		media_attachments: [],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109632442789543104',
-		created_at: '2023-01-04T18:51:58.940Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/lolgop/statuses/109632442789543104',
-		url: 'https://mastodon.social/@lolgop/109632442789543104',
-		replies_count: 5,
-		reblogs_count: 63,
-		favourites_count: 133,
-		edited_at: null,
-		content: '\u003cp\u003eImagine hating Kevin McCarthy because he isn\u0026#39;t terrible enough.\u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Mastodon Twitter Crossposter', website: null },
-		account: {
-			id: '109243448364080246',
-			username: 'lolgop',
-			acct: 'lolgop',
-			display_name: 'L O L G O P',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-10-28T00:00:00.000Z',
-			note: '\u003cp\u003eLess estranged from Elon than most of his kids.\u003c/p\u003e',
-			url: 'https://mastodon.social/@lolgop',
-			avatar: 'https://files.mastodon.social/accounts/avatars/109/243/448/364/080/246/original/c71af9ae6eb92594.jpg',
-			avatar_static:
-				'https://files.mastodon.social/accounts/avatars/109/243/448/364/080/246/original/c71af9ae6eb92594.jpg',
-			header: 'https://files.mastodon.social/accounts/headers/109/243/448/364/080/246/original/b4bbb0fa94cee3ff.jpg',
-			header_static:
-				'https://files.mastodon.social/accounts/headers/109/243/448/364/080/246/original/b4bbb0fa94cee3ff.jpg',
-			followers_count: 16877,
-			following_count: 151,
-			statuses_count: 1487,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [
-				{
-					name: 'Blue Walls',
-					value:
-						'\u003ca href="https://bluewalls.substack.com/" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003ebluewalls.substack.com/\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109630453680616888',
-		created_at: '2023-01-04T10:26:07.590Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/dmoser/statuses/109630453680616888',
-		url: 'https://mastodon.social/@dmoser/109630453680616888',
-		replies_count: 11,
-		reblogs_count: 287,
-		favourites_count: 437,
-		edited_at: null,
-		content:
-			'\u003cp\u003eParis 🇫🇷 is removing half its street parking to free up space for people. \u003c/p\u003e\u003cp\u003eCars 🚗 in Paris take up 50% ❗️of public space.👇👇\u003c/p\u003e\u003cp\u003e🎥 @wef@twitter.com\u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Mastodon Twitter Crossposter', website: null },
-		account: {
-			id: '109251842210716034',
-			username: 'dmoser',
-			acct: 'dmoser',
-			display_name: 'Daniel Moser',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-10-29T00:00:00.000Z',
-			note: '\u003cp\u003eSenior Transport Specialist @worldbank | Before: Head @tuminitiative | Mobility Transport | Urban Planning | Infrastructure | Government Advisory\u003c/p\u003e',
-			url: 'https://mastodon.social/@dmoser',
-			avatar: 'https://files.mastodon.social/accounts/avatars/109/251/842/210/716/034/original/c3cce22d3fa6154a.png',
-			avatar_static:
-				'https://files.mastodon.social/accounts/avatars/109/251/842/210/716/034/original/c3cce22d3fa6154a.png',
-			header: 'https://files.mastodon.social/accounts/headers/109/251/842/210/716/034/original/5d3ebd2b7eea6e7a.jpeg',
-			header_static:
-				'https://files.mastodon.social/accounts/headers/109/251/842/210/716/034/original/5d3ebd2b7eea6e7a.jpeg',
-			followers_count: 2832,
-			following_count: 96,
-			statuses_count: 234,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [
-				{
-					name: 'Twitter',
-					value:
-						'\u003ca href="https://mobile.twitter.com/_dmoser" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003emobile.twitter.com/_dmoser\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [
-			{
-				id: '109630453544447609',
-				type: 'video',
-				url: 'https://files.mastodon.social/media_attachments/files/109/630/453/544/447/609/original/9106c29b38b295ce.mp4',
-				preview_url:
-					'https://files.mastodon.social/media_attachments/files/109/630/453/544/447/609/small/9106c29b38b295ce.png',
-				remote_url: null,
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					original: { width: 720, height: 720, frame_rate: '24/1', duration: 76.352, bitrate: 762210 },
-					small: { width: 400, height: 400, size: '400x400', aspect: 1.0 },
-				},
-				description: null,
-				blurhash: 'UQFY.Lad9Et7_3xte,a~9FRjxuog00WB%Mj]',
-			},
-		],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109631141900228904',
-		created_at: '2023-01-04T13:21:08.938Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/nixCraft/statuses/109631141900228904',
-		url: 'https://mastodon.social/@nixCraft/109631141900228904',
-		replies_count: 10,
-		reblogs_count: 153,
-		favourites_count: 271,
-		edited_at: '2023-01-04T13:21:32.490Z',
-		content:
-			'\u003cp\u003eThe real message of Jurassic Park is that you get the Unix and IT support you pay for. Also, Jurassic Park is a cautionary tale about how you compensate your IT staff.\u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Web', website: null },
-		account: {
-			id: '38659',
-			username: 'nixCraft',
-			acct: 'nixCraft',
-			display_name: 'nixCraft 🐧',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2017-04-03T00:00:00.000Z',
-			note: '\u003cp\u003eEnjoy \u003ca href="https://mastodon.social/tags/Linux" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eLinux\u003c/span\u003e\u003c/a\u003e, \u003ca href="https://mastodon.social/tags/macOS" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003emacOS\u003c/span\u003e\u003c/a\u003e, \u003ca href="https://mastodon.social/tags/FreeBSD" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eFreeBSD\u003c/span\u003e\u003c/a\u003e \u0026amp; \u003ca href="https://mastodon.social/tags/Unix" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eUnix\u003c/span\u003e\u003c/a\u003e like systems? \u003ca href="https://mastodon.social/tags/Opensource" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eOpensource\u003c/span\u003e\u003c/a\u003e software \u0026amp; \u003ca href="https://mastodon.social/tags/programming" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eprogramming\u003c/span\u003e\u003c/a\u003e? Enjoy \u003ca href="https://mastodon.social/tags/Sysadmin" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eSysadmin\u003c/span\u003e\u003c/a\u003e \u0026amp; \u003ca href="https://mastodon.social/tags/DevOps" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eDevOps\u003c/span\u003e\u003c/a\u003e work? Follow us to make the most of your geeky IT career.\u003c/p\u003e',
-			url: 'https://mastodon.social/@nixCraft',
-			avatar: 'https://files.mastodon.social/accounts/avatars/000/038/659/original/b0de8058da52f2aa.jpg',
-			avatar_static: 'https://files.mastodon.social/accounts/avatars/000/038/659/original/b0de8058da52f2aa.jpg',
-			header: 'https://files.mastodon.social/accounts/headers/000/038/659/original/6c4d08cfd6c28163.png',
-			header_static: 'https://files.mastodon.social/accounts/headers/000/038/659/original/6c4d08cfd6c28163.png',
-			followers_count: 14151,
-			following_count: 13,
-			statuses_count: 824,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [
-				{
-					name: 'Linux blog',
-					value:
-						'\u003ca href="https://www.cyberciti.biz" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://www.\u003c/span\u003e\u003cspan class=""\u003ecyberciti.biz\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2022-11-21T11:10:53.864+00:00',
-				},
-				{
-					name: 'Q \u0026 A forum',
-					value:
-						'\u003ca href="https://www.nixcraft.com" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://www.\u003c/span\u003e\u003cspan class=""\u003enixcraft.com\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2022-12-18T08:41:11.287+00:00',
-				},
-			],
-		},
-		media_attachments: [
-			{
-				id: '109631143187061293',
-				type: 'image',
-				url: 'https://files.mastodon.social/media_attachments/files/109/631/143/187/061/293/original/2ae898b17f981c5e.png',
-				preview_url:
-					'https://files.mastodon.social/media_attachments/files/109/631/143/187/061/293/small/2ae898b17f981c5e.png',
-				remote_url: null,
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					original: { width: 640, height: 348, size: '640x348', aspect: 1.839080459770115 },
-					small: { width: 640, height: 348, size: '640x348', aspect: 1.839080459770115 },
-				},
-				description: null,
-				blurhash: 'UWF~U0D%4,-r9FohkEadWBtQahNF%gWAM_WY',
-			},
-		],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109632341701975927',
-		created_at: '2023-01-04T18:26:16.000Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://cyberplace.social/users/GossiTheDog/statuses/109632341711653832',
-		url: 'https://cyberplace.social/@GossiTheDog/109632341711653832',
-		replies_count: 11,
-		reblogs_count: 158,
-		favourites_count: 22,
-		edited_at: '2023-01-04T19:21:52.000Z',
-		content:
-			'\u003cp\u003eIf you’re outraged about Adobe sending your pictures off their servers (you should be), please know other vendors do this too. That horse has already bolted. \u003c/p\u003e\u003cp\u003eEg Microsoft Edge automatically sends your key presses in Edge to MS - enabled by default \u003ca href="https://techcommunity.microsoft.com/t5/discussions/how-to-disable-writing-assistance-via-group-policy/td-p/3648422" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class="ellipsis"\u003etechcommunity.microsoft.com/t5\u003c/span\u003e\u003cspan class="invisible"\u003e/discussions/how-to-disable-writing-assistance-via-group-policy/td-p/3648422\u003c/span\u003e\u003c/a\u003e\u003c/p\u003e\u003cp\u003eMicrosoft Office 365 sends every photo and screenshot you add in Word, PowerPoint etc (including in emails) to Microsoft 365 Intelligent Services without prompt \u003ca href="https://support.microsoft.com/en-us/office/everything-you-need-to-know-to-write-effective-alt-text-df98f884-ca3d-456c-807b-1a1fa82f5dc2" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class="ellipsis"\u003esupport.microsoft.com/en-us/of\u003c/span\u003e\u003cspan class="invisible"\u003efice/everything-you-need-to-know-to-write-effective-alt-text-df98f884-ca3d-456c-807b-1a1fa82f5dc2\u003c/span\u003e\u003c/a\u003e\u003c/p\u003e',
-		reblog: null,
-		account: {
-			id: '109387541307161560',
-			username: 'GossiTheDog',
-			acct: 'GossiTheDog@cyberplace.social',
-			display_name: 'Kevin Beaumont',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-11-22T00:00:00.000Z',
-			note: '\u003cp\u003eCybersecurity weather person and award winning shitposter. Shitposting is an anagram of Top Insights. \u003ca href="https://cyberplace.social/tags/fedi22" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003efedi22\u003c/span\u003e\u003c/a\u003e\u003c/p\u003e\u003cp\u003eI run the Mastodon server \u003ca href="https://cyberplace.social/tags/cyberplace" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003ecyberplace\u003c/span\u003e\u003c/a\u003e\u003c/p\u003e\u003cp\u003e\u003ca href="https://cyberplace.social/tags/cyber" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003ecyber\u003c/span\u003e\u003c/a\u003e | \u003ca href="https://cyberplace.social/tags/cybersecurity" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003ecybersecurity\u003c/span\u003e\u003c/a\u003e | \u003ca href="https://cyberplace.social/tags/infosec" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003einfosec\u003c/span\u003e\u003c/a\u003e | \u003ca href="https://cyberplace.social/tags/FlightSimulator" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003eFlightSimulator\u003c/span\u003e\u003c/a\u003e | \u003ca href="https://cyberplace.social/tags/tech" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003etech\u003c/span\u003e\u003c/a\u003e | \u003ca href="https://cyberplace.social/tags/ThreatIntel" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003eThreatIntel\u003c/span\u003e\u003c/a\u003e | \u003ca href="https://cyberplace.social/tags/AlwaysLearning" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003eAlwaysLearning\u003c/span\u003e\u003c/a\u003e | \u003ca href="https://cyberplace.social/tags/dogs" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003edogs\u003c/span\u003e\u003c/a\u003e | \u003ca href="https://cyberplace.social/tags/cats" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank"\u003e#\u003cspan\u003ecats\u003c/span\u003e\u003c/a\u003e\u003c/p\u003e',
-			url: 'https://cyberplace.social/@GossiTheDog',
-			avatar:
-				'https://files.mastodon.social/cache/accounts/avatars/109/387/541/307/161/560/original/89e70e03b2542f92.png',
-			avatar_static:
-				'https://files.mastodon.social/cache/accounts/avatars/109/387/541/307/161/560/original/89e70e03b2542f92.png',
-			header:
-				'https://files.mastodon.social/cache/accounts/headers/109/387/541/307/161/560/original/5d60ae8a8be2bc64.gif',
-			header_static:
-				'https://files.mastodon.social/cache/accounts/headers/109/387/541/307/161/560/static/5d60ae8a8be2bc64.png',
-			followers_count: 22017,
-			following_count: 701,
-			statuses_count: 1026,
-			last_status_at: '2023-01-04',
-			emojis: [],
-			fields: [
-				{
-					name: 'Website',
-					value:
-						'\u003ca href="https://doublepulsar.com" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003edoublepulsar.com\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Verified it me',
-					value:
-						'\u003ca href="https://twittodon.com/share.php?t=GossiTheDog\u0026amp;m=gossithedog@cyberplace.social" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class="ellipsis"\u003etwittodon.com/share.php?t=Goss\u003c/span\u003e\u003cspan class="invisible"\u003eiTheDog\u0026amp;m=gossithedog@cyberplace.social\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2023-01-04T19:53:03.077+00:00',
-				},
-				{
-					name: 'Twitter',
-					value:
-						'\u003ca href="https://fedified.com/@GossiTheDog" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003efedified.com/@GossiTheDog\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2023-01-04T19:53:03.132+00:00',
-				},
-				{
-					name: 'Cyberplace Status',
-					value:
-						'\u003ca href="http://status.cyberplace.social" rel="nofollow noopener noreferrer" target="_blank"\u003e\u003cspan class="invisible"\u003ehttp://\u003c/span\u003e\u003cspan class=""\u003estatus.cyberplace.social\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: {
-			url: 'https://techcommunity.microsoft.com/t5/discussions/how-to-disable-writing-assistance-via-group-policy/td-p/3648422',
-			title: 'How to disable writing assistance via Group Policy',
-			description:
-				'Hello,   new Edge version 106 has got settings "Use writing assistance". I look for matching settings in group policy for Edge but with not succes. I use the latest adml and admx files.  Group setting Disable spellcheck will help but only for Edge 104 (and lower). Edge 106 has replaced settings Spel...',
-			type: 'link',
-			author_name: '',
-			author_url: '',
-			provider_name: 'TECHCOMMUNITY.MICROSOFT.COM',
-			provider_url: '',
-			html: '',
-			width: 400,
-			height: 208,
-			image: 'https://files.mastodon.social/cache/preview_cards/images/051/386/707/original/fad1ee4a8d2bcea0.jpg',
-			embed_url: '',
-			blurhash: 'U76,iAhiy9O?zeygkbR4R-kUa8atD*M|kWtR',
-		},
-		poll: null,
-	},
-	{
-		id: '109632674024513398',
-		created_at: '2023-01-04T19:50:47.301Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/Popehat/statuses/109632674024513398',
-		url: 'https://mastodon.social/@Popehat/109632674024513398',
-		replies_count: 10,
-		reblogs_count: 37,
-		favourites_count: 112,
-		edited_at: null,
-		content:
-			'\u003cp\u003eJust recorded the first Serious Trouble episode of the new year, out tomorrow.  This week:  George Santos is in serious trouble.  Sam Bankman-Fried is in REALLY serious trouble.  And Scott Adams is still making dumb defamation threats.\u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Web', website: null },
-		account: {
-			id: '109502260753916593',
-			username: 'Popehat',
-			acct: 'Popehat',
-			display_name: 'Ken White',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-12-12T00:00:00.000Z',
-			note: '\u003cp\u003eCriminal defense attorney and First Amendment litigator.  Lawsplainer and podcaster.\u003c/p\u003e',
-			url: 'https://mastodon.social/@Popehat',
-			avatar: 'https://files.mastodon.social/accounts/avatars/109/502/260/753/916/593/original/f721da0f38083abf.jpg',
-			avatar_static:
-				'https://files.mastodon.social/accounts/avatars/109/502/260/753/916/593/original/f721da0f38083abf.jpg',
-			header: 'https://files.mastodon.social/accounts/headers/109/502/260/753/916/593/original/12414ec7db888f86.jpg',
-			header_static:
-				'https://files.mastodon.social/accounts/headers/109/502/260/753/916/593/original/12414ec7db888f86.jpg',
-			followers_count: 73092,
-			following_count: 248,
-			statuses_count: 282,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [
-				{
-					name: 'The Popehat Report',
-					value:
-						'\u003ca href="https://popehat.substack.com/" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003epopehat.substack.com/\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Serious Trouble Podcast',
-					value:
-						'\u003ca href="https://www.serioustrouble.show/" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://www.\u003c/span\u003e\u003cspan class=""\u003eserioustrouble.show/\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Popehat on Post.News',
-					value:
-						'\u003ca href="https://post.news/popehat" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003epost.news/popehat\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '109734788115874319',
-		created_at: '2023-01-22T20:39:44.678Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.design/users/rafa/statuses/109734788115874319',
-		url: 'https://mastodon.design/@rafa/109734788115874319',
-		replies_count: 43,
-		reblogs_count: 256,
-		favourites_count: 537,
-		edited_at: null,
-		content:
-			'\u003cp\u003eHi, meet HiDock!\u003c/p\u003e\u003cp\u003eIt\'s a free Mac app that lets you set different Dock settings for different display configurations\u003c/p\u003e\u003cp\u003e\u003ca href="https://hidock.app" target="_blank" rel="nofollow noopener noreferrer"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003ehidock.app\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e →\u003c/p\u003e',
-		reblog: null,
-		application: {
-			name: 'Web',
-			website: null,
-		},
-		account: {
-			id: '11932',
-			username: 'rafa',
-			acct: 'rafa',
-			display_name: 'Rafa',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2018-08-21T00:00:00.000Z',
-			note: '\u003cp\u003eI’m a designer and app developer, currently working on Sketch, and Hand Mirror for Mac\u003c/p\u003e',
-			url: 'https://mastodon.design/@rafa',
-			avatar: 'https://cdn.masto.host/mastodondesign/accounts/avatars/000/011/932/original/8f601be03c98b2e8.png',
-			avatar_static: 'https://cdn.masto.host/mastodondesign/accounts/avatars/000/011/932/original/8f601be03c98b2e8.png',
-			header: 'https://cdn.masto.host/mastodondesign/accounts/headers/000/011/932/original/668f6d32abb54252.jpeg',
-			header_static:
-				'https://cdn.masto.host/mastodondesign/accounts/headers/000/011/932/original/668f6d32abb54252.jpeg',
-			followers_count: 2511,
-			following_count: 437,
-			statuses_count: 1115,
-			last_status_at: '2023-01-23',
-			noindex: false,
-			emojis: [],
-			fields: [
-				{
-					name: 'Website',
-					value:
-						'\u003ca href="https://rafa.design" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://\u003c/span\u003e\u003cspan class=""\u003erafa.design\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2022-11-06T16:49:24.339+00:00',
-				},
-				{
-					name: 'App',
-					value:
-						'\u003ca href="http://handmirror.app" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttp://\u003c/span\u003e\u003cspan class=""\u003ehandmirror.app\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Podcast',
-					value:
-						'\u003ca href="http://layout.fm" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttp://\u003c/span\u003e\u003cspan class=""\u003elayout.fm\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: null,
-				},
-				{
-					name: 'Pronouns',
-					value: 'He/Them',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [
-			{
-				id: '109734774350200717',
-				type: 'image',
-				url: 'https://cdn.masto.host/mastodondesign/media_attachments/files/109/734/774/350/200/717/original/51054ec255c7a366.png',
-				preview_url:
-					'https://cdn.masto.host/mastodondesign/media_attachments/files/109/734/774/350/200/717/small/51054ec255c7a366.png',
-				remote_url: null,
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					original: {
-						width: 1821,
-						height: 1138,
-						size: '1821x1138',
-						aspect: 1.6001757469244289,
-					},
-					small: {
-						width: 606,
-						height: 379,
-						size: '606x379',
-						aspect: 1.5989445910290236,
-					},
-					focus: {
-						x: 0.0,
-						y: 0.0,
-					},
-				},
-				description:
-					'App icon for the app HiDock, against the default macOS Ventura background. The icon resembles an old school Apple display, with a aluminum frame, glossy black bezels, and what looks like the Snow Leopard wallpaper with a dock highlighted',
-				blurhash: 'UIQsS4?R}zRh%_$wKP-4-xt11mwwu0xUMmr?',
-			},
-			{
-				id: '109734775049506533',
-				type: 'image',
-				url: 'https://cdn.masto.host/mastodondesign/media_attachments/files/109/734/775/049/506/533/original/b6c22d380e50cd11.png',
-				preview_url:
-					'https://cdn.masto.host/mastodondesign/media_attachments/files/109/734/775/049/506/533/small/b6c22d380e50cd11.png',
-				remote_url: null,
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					original: {
-						width: 1821,
-						height: 1138,
-						size: '1821x1138',
-						aspect: 1.6001757469244289,
-					},
-					small: {
-						width: 606,
-						height: 379,
-						size: '606x379',
-						aspect: 1.5989445910290236,
-					},
-					focus: {
-						x: 0.0,
-						y: 0.0,
-					},
-				},
-				description:
-					'Screenshot of HiDock, and with a subtitle that reads "Set different Docks settings for different displays"',
-				blurhash: 'UaQ@*F%HqsadyWxVNGw[QkobPCo0%1n#Rls:',
-			},
-			{
-				id: '109734775691562296',
-				type: 'image',
-				url: 'https://cdn.masto.host/mastodondesign/media_attachments/files/109/734/775/691/562/296/original/266c1740f6e4cdc1.png',
-				preview_url:
-					'https://cdn.masto.host/mastodondesign/media_attachments/files/109/734/775/691/562/296/small/266c1740f6e4cdc1.png',
-				remote_url: null,
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					original: {
-						width: 1821,
-						height: 1138,
-						size: '1821x1138',
-						aspect: 1.6001757469244289,
-					},
-					small: {
-						width: 606,
-						height: 379,
-						size: '606x379',
-						aspect: 1.5989445910290236,
-					},
-					focus: {
-						x: 0.0,
-						y: 0.0,
-					},
-				},
-				description:
-					'Screenshot of HiDock, now in dark mode, with a caption that reads "Customize Size, Position, and Hidden Preference"',
-				blurhash: 'UgIB[UxW}Aox=Y$#xYs,sos-FdbFwJsmS4n+',
-			},
-			{
-				id: '109734776307839727',
-				type: 'image',
-				url: 'https://cdn.masto.host/mastodondesign/media_attachments/files/109/734/776/307/839/727/original/6085716d5f865ce1.png',
-				preview_url:
-					'https://cdn.masto.host/mastodondesign/media_attachments/files/109/734/776/307/839/727/small/6085716d5f865ce1.png',
-				remote_url: null,
-				preview_remote_url: null,
-				text_url: null,
-				meta: {
-					original: {
-						width: 1821,
-						height: 1138,
-						size: '1821x1138',
-						aspect: 1.6001757469244289,
-					},
-					small: {
-						width: 606,
-						height: 379,
-						size: '606x379',
-						aspect: 1.5989445910290236,
-					},
-					focus: {
-						x: 0.0,
-						y: 0.0,
-					},
-				},
-				description:
-					'Another screenshot of HiDock, this time you can see a "Meet the Developer section", with a caption that reads "Runs in the background, no clutter or extra icons"',
-				blurhash: 'UiH+IDxW}Toc-6xWxWs-xGs-F0bGwJsnS5n%',
-			},
-		],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: {
-			url: 'https://hidock.app/',
-			title: 'HiDock, for Mac',
-			description: 'Custom Dock settings for different displays',
-			type: 'link',
-			author_name: '',
-			author_url: '',
-			provider_name: 'HiDock, for Mac',
-			provider_url: '',
-			html: '',
-			width: 400,
-			height: 200,
-			image:
-				'https://cdn.masto.host/mastodondesign/cache/preview_cards/images/001/081/966/original/f9d41b35290d4ddf.jpg',
-			embed_url: '',
-			blurhash: 'UHIV*-iwxuIU0}mlBp9|V@gNiws:FxpIVsw^',
-		},
-		poll: null,
-	},
-	{
-		id: '109261798394272672',
-		created_at: '2022-10-31T07:52:14.290Z',
-		in_reply_to_id: null,
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.design/users/rafa/statuses/109261798394272672',
-		url: 'https://mastodon.design/@rafa/109261798394272672',
-		replies_count: 5,
-		reblogs_count: 10,
-		favourites_count: 42,
-		edited_at: null,
-		content:
-			'<p>Time for an updated <a href="https://mastodon.design/tags/Introduction" class="mention hashtag" rel="tag">#<span>Introduction</span></a>:</p><p>I&#39;m Rafa, a designer and app developer currently living in Amsterdam ☀️</p><p>I make Hand Mirror for macOS, Booby Track for iOS/watchOS, and co-host a design podcast with my friend <span class="h-card"><a href="https://mastodon.design/@kevin" class="u-url mention">@<span>kevin</span></a></span> called Layout.</p><p>Been loving the sense of community and the good vibes you&#39;re all giving in this platform, say hi 👋</p><p>Here&#39;s some topics I feel like I like to participate in, I hear hashtags can help with discovery so here it goes: <a href="https://mastodon.design/tags/Design" class="mention hashtag" rel="tag">#<span>Design</span></a> <a href="https://mastodon.design/tags/Tech" class="mention hashtag" rel="tag">#<span>Tech</span></a> <a href="https://mastodon.design/tags/AppDevelopment" class="mention hashtag" rel="tag">#<span>AppDevelopment</span></a> <a href="https://mastodon.design/tags/Parenthood" class="mention hashtag" rel="tag">#<span>Parenthood</span></a> <a href="https://mastodon.design/tags/Therapy" class="mention hashtag" rel="tag">#<span>Therapy</span></a></p>',
-		reblog: null,
-		application: {
-			name: 'Web',
-			website: null,
-		},
-		account: {
-			id: '11932',
-			username: 'rafa',
-			acct: 'rafa',
-			display_name: 'Rafa',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2018-08-21T00:00:00.000Z',
-			note: '<p>I’m a designer and app developer, currently working on Sketch, and Hand Mirror for Mac</p>',
-			url: 'https://mastodon.design/@rafa',
-			avatar: 'https://cdn.masto.host/mastodondesign/accounts/avatars/000/011/932/original/8f601be03c98b2e8.png',
-			avatar_static: 'https://cdn.masto.host/mastodondesign/accounts/avatars/000/011/932/original/8f601be03c98b2e8.png',
-			header: 'https://cdn.masto.host/mastodondesign/accounts/headers/000/011/932/original/668f6d32abb54252.jpeg',
-			header_static:
-				'https://cdn.masto.host/mastodondesign/accounts/headers/000/011/932/original/668f6d32abb54252.jpeg',
-			followers_count: 3135,
-			following_count: 453,
-			statuses_count: 1289,
-			last_status_at: '2023-02-12',
-			noindex: false,
-			emojis: [],
-			fields: [
-				{
-					name: 'Website',
-					value:
-						'<a href="https://rafa.design" target="_blank" rel="nofollow noopener noreferrer me"><span class="invisible">https://</span><span class="">rafa.design</span><span class="invisible"></span></a>',
-					verified_at: '2022-11-06T16:49:24.339+00:00',
-				},
-				{
-					name: 'App',
-					value:
-						'<a href="http://handmirror.app" target="_blank" rel="nofollow noopener noreferrer me"><span class="invisible">http://</span><span class="">handmirror.app</span><span class="invisible"></span></a>',
-					verified_at: null,
-				},
-				{
-					name: 'Podcast',
-					value:
-						'<a href="http://layout.fm" target="_blank" rel="nofollow noopener noreferrer me"><span class="invisible">http://</span><span class="">layout.fm</span><span class="invisible"></span></a>',
-					verified_at: null,
-				},
-				{
-					name: 'Pronouns',
-					value: 'He/Them',
-					verified_at: null,
-				},
-			],
-		},
-		media_attachments: [],
-		mentions: [
-			{
-				id: '12179',
-				username: 'kevin',
-				url: 'https://mastodon.design/@kevin',
-				acct: 'kevin',
-			},
-		],
-		tags: [
-			{
-				name: 'therapy',
-				url: 'https://mastodon.design/tags/therapy',
-			},
-			{
-				name: 'parenthood',
-				url: 'https://mastodon.design/tags/parenthood',
-			},
-			{
-				name: 'appdevelopment',
-				url: 'https://mastodon.design/tags/appdevelopment',
-			},
-			{
-				name: 'tech',
-				url: 'https://mastodon.design/tags/tech',
-			},
-			{
-				name: 'design',
-				url: 'https://mastodon.design/tags/design',
-			},
-			{
-				name: 'introduction',
-				url: 'https://mastodon.design/tags/introduction',
-			},
-		],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-]
-
-export const statuses: MastodonStatus[] = mastodonRawStatuses.map((rawStatus) => ({
-	...rawStatus,
-	media_attachments: rawStatus.media_attachments.map((mediaAttachment) => ({
-		...mediaAttachment,
-		type: getStandardMediaType(mediaAttachment.type),
-	})),
-}))
-
-function getStandardMediaType(mediaAttachmentMastodonType: string): string {
-	switch (mediaAttachmentMastodonType) {
-		case 'image':
-			return 'Image'
-		case 'video':
-			return 'Video'
-	}
-	return mediaAttachmentMastodonType
-}
-
-export const replies: MastodonStatus[] = [
-	{
-		id: '209630407170172986',
-		created_at: '2023-01-04T17:14:17.855Z',
-		in_reply_to_id: '109630407170172986',
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/ZachWeinersmith/statuses/209630407170172986',
-		url: 'https://mastodon.social/@ZachWeinersmith/209630407170172986',
-		replies_count: 52,
-		reblogs_count: 630,
-		favourites_count: 1434,
-		edited_at: '2023-01-04T10:15:59.795Z',
-		content: '\u003cp\u003e Yes we did! \u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Web', website: null },
-		account: {
-			id: '109521032613939160',
-			username: 'ZachWeinersmith',
-			acct: 'ZachWeinersmith',
-			display_name: 'Zach Weinersmith',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2022-12-16T00:00:00.000Z',
-			note: '\u003cp\u003eThe SMBC guy. Co-author of Soonish, illustrator of Open Borders, scop of Bea Wolf.\u003c/p\u003e',
-			url: 'https://mastodon.social/@ZachWeinersmith',
-			avatar: 'https://files.mastodon.social/accounts/avatars/109/521/032/613/939/160/original/6d6588cad807dfbc.png',
-			avatar_static:
-				'https://files.mastodon.social/accounts/avatars/109/521/032/613/939/160/original/6d6588cad807dfbc.png',
-			header: 'https://files.mastodon.social/accounts/headers/109/521/032/613/939/160/original/8368e47ebdb7f00f.jpg',
-			header_static:
-				'https://files.mastodon.social/accounts/headers/109/521/032/613/939/160/original/8368e47ebdb7f00f.jpg',
-			followers_count: 4112,
-			following_count: 514,
-			statuses_count: 142,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [],
-		},
-		media_attachments: [],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-	{
-		id: '309630407170172986',
-		created_at: '2023-01-04T22:14:17.855Z',
-		in_reply_to_id: '109630407170172986',
-		in_reply_to_account_id: null,
-		sensitive: false,
-		spoiler_text: '',
-		visibility: 'public',
-		language: 'en',
-		uri: 'https://mastodon.social/users/ZachWeinersmith/statuses/209630407170172986',
-		url: 'https://mastodon.social/@ZachWeinersmith/209630407170172986',
-		replies_count: 52,
-		reblogs_count: 630,
-		favourites_count: 1434,
-		edited_at: '2023-01-04T10:15:59.795Z',
-		content: '\u003cp\u003e Yes you guys did it! \u003c/p\u003e',
-		reblog: null,
-		application: { name: 'Web', website: null },
-		account: {
-			id: '38659',
-			username: 'nixCraft',
-			acct: 'nixCraft',
-			display_name: 'nixCraft 🐧',
-			locked: false,
-			bot: false,
-			discoverable: true,
-			group: false,
-			created_at: '2017-04-03T00:00:00.000Z',
-			note: '\u003cp\u003eEnjoy \u003ca href="https://mastodon.social/tags/Linux" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eLinux\u003c/span\u003e\u003c/a\u003e, \u003ca href="https://mastodon.social/tags/macOS" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003emacOS\u003c/span\u003e\u003c/a\u003e, \u003ca href="https://mastodon.social/tags/FreeBSD" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eFreeBSD\u003c/span\u003e\u003c/a\u003e \u0026amp; \u003ca href="https://mastodon.social/tags/Unix" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eUnix\u003c/span\u003e\u003c/a\u003e like systems? \u003ca href="https://mastodon.social/tags/Opensource" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eOpensource\u003c/span\u003e\u003c/a\u003e software \u0026amp; \u003ca href="https://mastodon.social/tags/programming" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eprogramming\u003c/span\u003e\u003c/a\u003e? Enjoy \u003ca href="https://mastodon.social/tags/Sysadmin" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eSysadmin\u003c/span\u003e\u003c/a\u003e \u0026amp; \u003ca href="https://mastodon.social/tags/DevOps" class="mention hashtag" rel="tag"\u003e#\u003cspan\u003eDevOps\u003c/span\u003e\u003c/a\u003e work? Follow us to make the most of your geeky IT career.\u003c/p\u003e',
-			url: 'https://mastodon.social/@nixCraft',
-			avatar: 'https://files.mastodon.social/accounts/avatars/000/038/659/original/b0de8058da52f2aa.jpg',
-			avatar_static: 'https://files.mastodon.social/accounts/avatars/000/038/659/original/b0de8058da52f2aa.jpg',
-			header: 'https://files.mastodon.social/accounts/headers/000/038/659/original/6c4d08cfd6c28163.png',
-			header_static: 'https://files.mastodon.social/accounts/headers/000/038/659/original/6c4d08cfd6c28163.png',
-			followers_count: 14151,
-			following_count: 13,
-			statuses_count: 824,
-			last_status_at: '2023-01-04',
-			noindex: false,
-			emojis: [],
-			fields: [
-				{
-					name: 'Linux blog',
-					value:
-						'\u003ca href="https://www.cyberciti.biz" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://www.\u003c/span\u003e\u003cspan class=""\u003ecyberciti.biz\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2022-11-21T11:10:53.864+00:00',
-				},
-				{
-					name: 'Q \u0026 A forum',
-					value:
-						'\u003ca href="https://www.nixcraft.com" target="_blank" rel="nofollow noopener noreferrer me"\u003e\u003cspan class="invisible"\u003ehttps://www.\u003c/span\u003e\u003cspan class=""\u003enixcraft.com\u003c/span\u003e\u003cspan class="invisible"\u003e\u003c/span\u003e\u003c/a\u003e',
-					verified_at: '2022-12-18T08:41:11.287+00:00',
-				},
-			],
-		},
-		media_attachments: [],
-		mentions: [],
-		tags: [],
-		emojis: [],
-		card: null,
-		poll: null,
-	},
-]
-
-export const tags: TagDetails[] = [
-	{
-		name: 'wintersolstice',
-		url: 'https://mastodon.social/tags/wintersolstice',
-		history: [
-			{ day: '1671580800', accounts: '951', uses: '1113' },
-			{ day: '1671494400', accounts: '59', uses: '75' },
-			{ day: '1671408000', accounts: '24', uses: '25' },
-			{ day: '1671321600', accounts: '19', uses: '21' },
-			{ day: '1671235200', accounts: '13', uses: '16' },
-			{ day: '1671148800', accounts: '13', uses: '13' },
-			{ day: '1671062400', accounts: '8', uses: '8' },
-		],
-	},
-	{
-		name: 'solstice',
-		url: 'https://mastodon.social/tags/solstice',
-		history: [
-			{ day: '1671580800', accounts: '954', uses: '1114' },
-			{ day: '1671494400', accounts: '80', uses: '118' },
-			{ day: '1671408000', accounts: '19', uses: '22' },
-			{ day: '1671321600', accounts: '24', uses: '28' },
-			{ day: '1671235200', accounts: '14', uses: '14' },
-			{ day: '1671148800', accounts: '9', uses: '9' },
-			{ day: '1671062400', accounts: '11', uses: '11' },
-		],
-	},
-	{
-		name: 'transitiontuesday',
-		url: 'https://mastodon.social/tags/transitiontuesday',
-		history: [
-			{ day: '1671580800', accounts: '92', uses: '103' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-			{ day: '1671148800', accounts: '0', uses: '0' },
-			{ day: '1671062400', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		name: 'mustardmovies',
-		url: 'https://mastodon.social/tags/mustardmovies',
-		history: [
-			{ day: '1671580800', accounts: '88', uses: '437' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-			{ day: '1671148800', accounts: '0', uses: '0' },
-			{ day: '1671062400', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		name: 'oddesttrumptaxdeductions',
-		url: 'https://mastodon.social/tags/oddesttrumptaxdeductions',
-		history: [
-			{ day: '1671580800', accounts: '73', uses: '113' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-			{ day: '1671148800', accounts: '0', uses: '0' },
-			{ day: '1671062400', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		name: 'wintersonnenwende',
-		url: 'https://mastodon.social/tags/wintersonnenwende',
-		history: [
-			{ day: '1671580800', accounts: '162', uses: '176' },
-			{ day: '1671494400', accounts: '8', uses: '8' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '1', uses: '1' },
-			{ day: '1671235200', accounts: '2', uses: '2' },
-			{ day: '1671148800', accounts: '0', uses: '0' },
-			{ day: '1671062400', accounts: '1', uses: '1' },
-		],
-	},
-	{
-		name: 'solsticehedgehog',
-		url: 'https://mastodon.social/tags/solsticehedgehog',
-		history: [
-			{ day: '1671580800', accounts: '54', uses: '61' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-			{ day: '1671148800', accounts: '0', uses: '0' },
-			{ day: '1671062400', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		name: 'waterfallwednesday',
-		url: 'https://mastodon.social/tags/waterfallwednesday',
-		history: [
-			{ day: '1671580800', accounts: '109', uses: '116' },
-			{ day: '1671494400', accounts: '4', uses: '4' },
-			{ day: '1671408000', accounts: '3', uses: '3' },
-			{ day: '1671321600', accounts: '1', uses: '1' },
-			{ day: '1671235200', accounts: '2', uses: '2' },
-			{ day: '1671148800', accounts: '5', uses: '6' },
-			{ day: '1671062400', accounts: '77', uses: '82' },
-		],
-	},
-	{
-		name: 'bautzen',
-		url: 'https://mastodon.social/tags/bautzen',
-		history: [
-			{ day: '1671580800', accounts: '50', uses: '62' },
-			{ day: '1671494400', accounts: '1', uses: '2' },
-			{ day: '1671408000', accounts: '5', uses: '11' },
-			{ day: '1671321600', accounts: '1', uses: '1' },
-			{ day: '1671235200', accounts: '2', uses: '2' },
-			{ day: '1671148800', accounts: '7', uses: '8' },
-			{ day: '1671062400', accounts: '9', uses: '13' },
-		],
-	},
-	{
-		name: 'googlefonts',
-		url: 'https://mastodon.social/tags/googlefonts',
-		history: [
-			{ day: '1671580800', accounts: '47', uses: '47' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '1', uses: '1' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-			{ day: '1671148800', accounts: '0', uses: '0' },
-			{ day: '1671062400', accounts: '1', uses: '1' },
-		],
-	},
-]
-
-export const links: Array<MastodonLink> = [
-	{
-		url: 'https://www.forbes.com/sites/emilybaker-white/2022/12/22/tiktok-tracks-forbes-journalists-bytedance/',
-		title: 'EXCLUSIVE: TikTok Spied On Forbes Journalists',
-		description:
-			'ByteDance confirmed it used TikTok to monitor journalists’ physical location using their IP addresses, as first reported by Forbes in October.',
-		type: 'link',
-		author_name: 'Emily Baker-White',
-		author_url: 'https://www.forbes.com/sites/emilybaker-white/',
-		provider_name: 'Forbes',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 161,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/524/205/original/471cd0adc1ee6409.jpg',
-		embed_url: '',
-		blurhash: 'UICsHx$*0Ko}w{RPnix]E1NZ-;s:gNkWRjM{',
-		history: [
-			{ day: '1671753600', accounts: '1503', uses: '1657' },
-			{ day: '1671667200', accounts: '1537', uses: '1727' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://www.theverge.com/2022/12/22/23523322/lastpass-data-breach-cloud-encrypted-password-vault-hackers',
-		title: 'Hackers stole encrypted LastPass password vaults, and we’re just now hearing about it',
-		description:
-			'LastPass has announced that during a November data breach of its cloud storage, hackers copied a backup of customer vault data that includes encrypted usernames and passwords.',
-		type: 'link',
-		author_name: 'Mitchell Clark',
-		author_url: '',
-		provider_name: 'The Verge',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 209,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/539/308/original/64c8d7b28cfaad2f.jpg',
-		embed_url: '',
-		blurhash: 'UfF;eEs,FRSiD^ajbXWXtvbb#hslECWVs%oL',
-		history: [
-			{ day: '1671753600', accounts: '282', uses: '291' },
-			{ day: '1671667200', accounts: '0', uses: '0' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://arstechnica.com/tech-policy/2022/12/facial-recognition-flags-girl-scout-mom-as-security-risk-at-rockettes-show/',
-		title: 'MSG defends using facial recognition to kick lawyer out of Rockettes show',
-		description: 'MSG Entertainment began using facial recognition at venues in 2018.',
-		type: 'link',
-		author_name: '',
-		author_url: '',
-		provider_name: 'Ars Technica',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 267,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/356/463/original/3d7c9f4eb7630b3a.jpg',
-		embed_url: '',
-		blurhash: 'ULBV@rQ+R%krPWR4n$o#1JV@oJof#RofbHWB',
-		history: [
-			{ day: '1671753600', accounts: '704', uses: '705' },
-			{ day: '1671667200', accounts: '3191', uses: '3228' },
-			{ day: '1671580800', accounts: '48', uses: '51' },
-			{ day: '1671494400', accounts: '17', uses: '17' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://www.theatlantic.com/ideas/archive/2022/12/zelensky-congress-speech-us-ukraine-support/672547/',
-		title: 'The Brutal Alternate World in Which the U.S. Abandoned Ukraine',
-		description: 'Ukrainian resistance and American support prevented a wide range of horrors.',
-		type: 'link',
-		author_name: 'Anne Applebaum',
-		author_url: '',
-		provider_name: 'The Atlantic',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 208,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/495/789/original/a464705b4eba4614.jpg',
-		embed_url: '',
-		blurhash: 'U14B:,4oW9M_Roavt8ax4nxt%NRjRhohayax',
-		history: [
-			{ day: '1671753600', accounts: '113', uses: '117' },
-			{ day: '1671667200', accounts: '441', uses: '452' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://www.washingtonpost.com/national-security/2022/12/22/jan-6-committee-trump-should-never-hold-office-again/',
-		title: 'Jan. 6 committee: Trump should never hold office again',
-		description:
-			'The committee’s recommendation came as part of an 800-plus page report that marks the culmination of its 18-month investigation',
-		type: 'link',
-		author_name: 'Amy Gardner',
-		author_url: 'https://www.washingtonpost.com/people/amy-gardner/',
-		provider_name: 'The Washington Post',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 267,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/548/842/original/5a159d17dc4b084a.jpeg',
-		embed_url: '',
-		blurhash: 'UKE.kHxaI9RO~TofMwn#SiWXNHWBIAWVtSoz',
-		history: [
-			{ day: '1671753600', accounts: '116', uses: '119' },
-			{ day: '1671667200', accounts: '0', uses: '0' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://www.washingtonpost.com/media/2022/12/23/musk-twitter-journalists-suspended-elonjet/',
-		title: 'Journalists who won’t delete Musk tweets remain locked out of Twitter',
-		description:
-			'Musk suspended reporters from Twitter and later reinstated them, with a catch: They must delete tweets related to the account @ElonJet, which has tracked Musk’s plane using public data.',
-		type: 'link',
-		author_name: 'Paul Farhi',
-		author_url: 'https://www.washingtonpost.com/people/paul-farhi/',
-		provider_name: 'The Washington Post',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 267,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/596/043/original/699238391e3433a5.jpeg',
-		embed_url: '',
-		blurhash: 'UA8gjY9DEMJDxUjdInNeIoNG-p$y9tt7xaxZ',
-		history: [
-			{ day: '1671753600', accounts: '119', uses: '120' },
-			{ day: '1671667200', accounts: '0', uses: '0' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://www.reuters.com/legal/facebook-parent-meta-pay-725-mln-settle-lawsuit-relating-cambridge-analytica-2022-12-23/',
-		title: 'Facebook parent Meta to settle Cambridge Analytica scandal case for $725 mln',
-		description:
-			'Facebook owner Meta Platforms Inc \u003ca href="https://www.reuters.com/companies/META.O" target="_blank"\u003e(META.O)\u003c/a\u003e has agreed to pay $725 million to resolve a class-action lawsuit accusing the social media giant of allowing third parties, including Cambridge Analytica, to access users\' personal information.',
-		type: 'link',
-		author_name: '',
-		author_url: '',
-		provider_name: 'Reuters',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 209,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/566/114/original/1d0630d4cf841790.jpg',
-		embed_url: '',
-		blurhash: 'UPNKbeMwMH?GWBWAog.8McofR*%NNHIUof-;',
-		history: [
-			{ day: '1671753600', accounts: '92', uses: '97' },
-			{ day: '1671667200', accounts: '0', uses: '0' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://www.bloomberg.com/news/articles/2022-12-22/musk-s-frequent-twitter-polls-are-at-risk-of-bot-manipulation',
-		title: 'Musk’s Frequent Twitter Polls Are at Risk of Bot Manipulation',
-		description: 'New research shows votes can be easily purchased during Twitter polls',
-		type: 'link',
-		author_name: 'Davey Alba',
-		author_url: '',
-		provider_name: 'Bloomberg',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 267,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/514/000/original/1bf993cf6612abd8.jpg',
-		embed_url: '',
-		blurhash: 'UHAfCIofo~o#ysWYeRWCDibHWAj?t8j@o}of',
-		history: [
-			{ day: '1671753600', accounts: '23', uses: '24' },
-			{ day: '1671667200', accounts: '170', uses: '196' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://www.wsj.com/articles/u-s-life-expectancy-fell-to-lowest-level-since-1996-11671667059',
-		title: 'U.S. Life Expectancy Fell to Lowest Level Since 1996',
-		description: 'Covid-19 and opioid overdoses contributed to a 5% rise in death rate last year',
-		type: 'link',
-		author_name: 'Julie Wernau',
-		author_url: '',
-		provider_name: 'The Wall Street Journal',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 200,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/485/157/original/3668ba0bbaf09ebc.jpeg',
-		embed_url: '',
-		blurhash: 'UkMDZ69^v$}qrYkCo|M}OYw^I:OXRjj[t7jF',
-		history: [
-			{ day: '1671753600', accounts: '31', uses: '31' },
-			{ day: '1671667200', accounts: '165', uses: '168' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-	{
-		url: 'https://www.wired.com/story/email-scam-dicks-sporting-goods-yeti-cooler/',
-		title: 'No, You Haven’t Won a Yeti Cooler From Dick’s Sporting Goods',
-		description: 'The future of email spam utilizes a coding trick that evades the most sophisticated detection tools.',
-		type: 'link',
-		author_name: 'Lauren Goode',
-		author_url: '',
-		provider_name: 'WIRED',
-		provider_url: '',
-		html: '',
-		width: 400,
-		height: 209,
-		image: 'https://files.mastodon.social/cache/preview_cards/images/050/574/335/original/e10dbfa75b92a64a.jpg',
-		embed_url: '',
-		blurhash: 'UMBVv7#GI_OTV$-Lx9V|5eSw$dxDtJEoJEX7',
-		history: [
-			{ day: '1671753600', accounts: '70', uses: '72' },
-			{ day: '1671667200', accounts: '0', uses: '0' },
-			{ day: '1671580800', accounts: '0', uses: '0' },
-			{ day: '1671494400', accounts: '0', uses: '0' },
-			{ day: '1671408000', accounts: '0', uses: '0' },
-			{ day: '1671321600', accounts: '0', uses: '0' },
-			{ day: '1671235200', accounts: '0', uses: '0' },
-		],
-	},
-]
