+++ frontend/src/dummyData/index.tsx
@@ -1,0 +1,3 @@
+export * from './statuses'
+export * from './tags'
+export * from './links'
