+++ frontend/src/dummyData/links.ts
@@ -1,0 +1,259 @@
+import { MastodonLink } from '../types'
+
+export const links: Array<MastodonLink> = [
+	{
+		url: 'https://www.forbes.com/sites/emilybaker-white/2022/12/22/tiktok-tracks-forbes-journalists-bytedance/',
+		title: 'EXCLUSIVE: TikTok Spied On Forbes Journalists',
+		description:
+			'ByteDance confirmed it used TikTok to monitor journalists’ physical location using their IP addresses, as first reported by Forbes in October.',
+		type: 'link',
+		author_name: 'Emily Baker-White',
+		author_url: 'https://www.forbes.com/sites/emilybaker-white/',
+		provider_name: 'Forbes',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 161,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/524/205/original/471cd0adc1ee6409.jpg',
+		embed_url: '',
+		blurhash: 'UICsHx$*0Ko}w{RPnix]E1NZ-;s:gNkWRjM{',
+		history: [
+			{ day: '1671753600', accounts: '1503', uses: '1657' },
+			{ day: '1671667200', accounts: '1537', uses: '1727' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://www.theverge.com/2022/12/22/23523322/lastpass-data-breach-cloud-encrypted-password-vault-hackers',
+		title: 'Hackers stole encrypted LastPass password vaults, and we’re just now hearing about it',
+		description:
+			'LastPass has announced that during a November data breach of its cloud storage, hackers copied a backup of customer vault data that includes encrypted usernames and passwords.',
+		type: 'link',
+		author_name: 'Mitchell Clark',
+		author_url: '',
+		provider_name: 'The Verge',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 209,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/539/308/original/64c8d7b28cfaad2f.jpg',
+		embed_url: '',
+		blurhash: 'UfF;eEs,FRSiD^ajbXWXtvbb#hslECWVs%oL',
+		history: [
+			{ day: '1671753600', accounts: '282', uses: '291' },
+			{ day: '1671667200', accounts: '0', uses: '0' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://arstechnica.com/tech-policy/2022/12/facial-recognition-flags-girl-scout-mom-as-security-risk-at-rockettes-show/',
+		title: 'MSG defends using facial recognition to kick lawyer out of Rockettes show',
+		description: 'MSG Entertainment began using facial recognition at venues in 2018.',
+		type: 'link',
+		author_name: '',
+		author_url: '',
+		provider_name: 'Ars Technica',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 267,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/356/463/original/3d7c9f4eb7630b3a.jpg',
+		embed_url: '',
+		blurhash: 'ULBV@rQ+R%krPWR4n$o#1JV@oJof#RofbHWB',
+		history: [
+			{ day: '1671753600', accounts: '704', uses: '705' },
+			{ day: '1671667200', accounts: '3191', uses: '3228' },
+			{ day: '1671580800', accounts: '48', uses: '51' },
+			{ day: '1671494400', accounts: '17', uses: '17' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://www.theatlantic.com/ideas/archive/2022/12/zelensky-congress-speech-us-ukraine-support/672547/',
+		title: 'The Brutal Alternate World in Which the U.S. Abandoned Ukraine',
+		description: 'Ukrainian resistance and American support prevented a wide range of horrors.',
+		type: 'link',
+		author_name: 'Anne Applebaum',
+		author_url: '',
+		provider_name: 'The Atlantic',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 208,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/495/789/original/a464705b4eba4614.jpg',
+		embed_url: '',
+		blurhash: 'U14B:,4oW9M_Roavt8ax4nxt%NRjRhohayax',
+		history: [
+			{ day: '1671753600', accounts: '113', uses: '117' },
+			{ day: '1671667200', accounts: '441', uses: '452' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://www.washingtonpost.com/national-security/2022/12/22/jan-6-committee-trump-should-never-hold-office-again/',
+		title: 'Jan. 6 committee: Trump should never hold office again',
+		description:
+			'The committee’s recommendation came as part of an 800-plus page report that marks the culmination of its 18-month investigation',
+		type: 'link',
+		author_name: 'Amy Gardner',
+		author_url: 'https://www.washingtonpost.com/people/amy-gardner/',
+		provider_name: 'The Washington Post',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 267,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/548/842/original/5a159d17dc4b084a.jpeg',
+		embed_url: '',
+		blurhash: 'UKE.kHxaI9RO~TofMwn#SiWXNHWBIAWVtSoz',
+		history: [
+			{ day: '1671753600', accounts: '116', uses: '119' },
+			{ day: '1671667200', accounts: '0', uses: '0' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://www.washingtonpost.com/media/2022/12/23/musk-twitter-journalists-suspended-elonjet/',
+		title: 'Journalists who won’t delete Musk tweets remain locked out of Twitter',
+		description:
+			'Musk suspended reporters from Twitter and later reinstated them, with a catch: They must delete tweets related to the account @ElonJet, which has tracked Musk’s plane using public data.',
+		type: 'link',
+		author_name: 'Paul Farhi',
+		author_url: 'https://www.washingtonpost.com/people/paul-farhi/',
+		provider_name: 'The Washington Post',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 267,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/596/043/original/699238391e3433a5.jpeg',
+		embed_url: '',
+		blurhash: 'UA8gjY9DEMJDxUjdInNeIoNG-p$y9tt7xaxZ',
+		history: [
+			{ day: '1671753600', accounts: '119', uses: '120' },
+			{ day: '1671667200', accounts: '0', uses: '0' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://www.reuters.com/legal/facebook-parent-meta-pay-725-mln-settle-lawsuit-relating-cambridge-analytica-2022-12-23/',
+		title: 'Facebook parent Meta to settle Cambridge Analytica scandal case for $725 mln',
+		description:
+			'Facebook owner Meta Platforms Inc \u003ca href="https://www.reuters.com/companies/META.O" target="_blank"\u003e(META.O)\u003c/a\u003e has agreed to pay $725 million to resolve a class-action lawsuit accusing the social media giant of allowing third parties, including Cambridge Analytica, to access users\' personal information.',
+		type: 'link',
+		author_name: '',
+		author_url: '',
+		provider_name: 'Reuters',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 209,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/566/114/original/1d0630d4cf841790.jpg',
+		embed_url: '',
+		blurhash: 'UPNKbeMwMH?GWBWAog.8McofR*%NNHIUof-;',
+		history: [
+			{ day: '1671753600', accounts: '92', uses: '97' },
+			{ day: '1671667200', accounts: '0', uses: '0' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://www.bloomberg.com/news/articles/2022-12-22/musk-s-frequent-twitter-polls-are-at-risk-of-bot-manipulation',
+		title: 'Musk’s Frequent Twitter Polls Are at Risk of Bot Manipulation',
+		description: 'New research shows votes can be easily purchased during Twitter polls',
+		type: 'link',
+		author_name: 'Davey Alba',
+		author_url: '',
+		provider_name: 'Bloomberg',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 267,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/514/000/original/1bf993cf6612abd8.jpg',
+		embed_url: '',
+		blurhash: 'UHAfCIofo~o#ysWYeRWCDibHWAj?t8j@o}of',
+		history: [
+			{ day: '1671753600', accounts: '23', uses: '24' },
+			{ day: '1671667200', accounts: '170', uses: '196' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://www.wsj.com/articles/u-s-life-expectancy-fell-to-lowest-level-since-1996-11671667059',
+		title: 'U.S. Life Expectancy Fell to Lowest Level Since 1996',
+		description: 'Covid-19 and opioid overdoses contributed to a 5% rise in death rate last year',
+		type: 'link',
+		author_name: 'Julie Wernau',
+		author_url: '',
+		provider_name: 'The Wall Street Journal',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 200,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/485/157/original/3668ba0bbaf09ebc.jpeg',
+		embed_url: '',
+		blurhash: 'UkMDZ69^v$}qrYkCo|M}OYw^I:OXRjj[t7jF',
+		history: [
+			{ day: '1671753600', accounts: '31', uses: '31' },
+			{ day: '1671667200', accounts: '165', uses: '168' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		url: 'https://www.wired.com/story/email-scam-dicks-sporting-goods-yeti-cooler/',
+		title: 'No, You Haven’t Won a Yeti Cooler From Dick’s Sporting Goods',
+		description: 'The future of email spam utilizes a coding trick that evades the most sophisticated detection tools.',
+		type: 'link',
+		author_name: 'Lauren Goode',
+		author_url: '',
+		provider_name: 'WIRED',
+		provider_url: '',
+		html: '',
+		width: 400,
+		height: 209,
+		image: 'https://files.mastodon.social/cache/preview_cards/images/050/574/335/original/e10dbfa75b92a64a.jpg',
+		embed_url: '',
+		blurhash: 'UMBVv7#GI_OTV$-Lx9V|5eSw$dxDtJEoJEX7',
+		history: [
+			{ day: '1671753600', accounts: '70', uses: '72' },
+			{ day: '1671667200', accounts: '0', uses: '0' },
+			{ day: '1671580800', accounts: '0', uses: '0' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+		],
+	},
+]
