+++ frontend/src/dummyData/tags.ts
@@ -1,0 +1,134 @@
+import { type TagDetails } from '../types'
+
+export const tags: TagDetails[] = [
+	{
+		name: 'wintersolstice',
+		url: 'https://mastodon.social/tags/wintersolstice',
+		history: [
+			{ day: '1671580800', accounts: '951', uses: '1113' },
+			{ day: '1671494400', accounts: '59', uses: '75' },
+			{ day: '1671408000', accounts: '24', uses: '25' },
+			{ day: '1671321600', accounts: '19', uses: '21' },
+			{ day: '1671235200', accounts: '13', uses: '16' },
+			{ day: '1671148800', accounts: '13', uses: '13' },
+			{ day: '1671062400', accounts: '8', uses: '8' },
+		],
+	},
+	{
+		name: 'solstice',
+		url: 'https://mastodon.social/tags/solstice',
+		history: [
+			{ day: '1671580800', accounts: '954', uses: '1114' },
+			{ day: '1671494400', accounts: '80', uses: '118' },
+			{ day: '1671408000', accounts: '19', uses: '22' },
+			{ day: '1671321600', accounts: '24', uses: '28' },
+			{ day: '1671235200', accounts: '14', uses: '14' },
+			{ day: '1671148800', accounts: '9', uses: '9' },
+			{ day: '1671062400', accounts: '11', uses: '11' },
+		],
+	},
+	{
+		name: 'transitiontuesday',
+		url: 'https://mastodon.social/tags/transitiontuesday',
+		history: [
+			{ day: '1671580800', accounts: '92', uses: '103' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+			{ day: '1671148800', accounts: '0', uses: '0' },
+			{ day: '1671062400', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		name: 'mustardmovies',
+		url: 'https://mastodon.social/tags/mustardmovies',
+		history: [
+			{ day: '1671580800', accounts: '88', uses: '437' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+			{ day: '1671148800', accounts: '0', uses: '0' },
+			{ day: '1671062400', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		name: 'oddesttrumptaxdeductions',
+		url: 'https://mastodon.social/tags/oddesttrumptaxdeductions',
+		history: [
+			{ day: '1671580800', accounts: '73', uses: '113' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+			{ day: '1671148800', accounts: '0', uses: '0' },
+			{ day: '1671062400', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		name: 'wintersonnenwende',
+		url: 'https://mastodon.social/tags/wintersonnenwende',
+		history: [
+			{ day: '1671580800', accounts: '162', uses: '176' },
+			{ day: '1671494400', accounts: '8', uses: '8' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '1', uses: '1' },
+			{ day: '1671235200', accounts: '2', uses: '2' },
+			{ day: '1671148800', accounts: '0', uses: '0' },
+			{ day: '1671062400', accounts: '1', uses: '1' },
+		],
+	},
+	{
+		name: 'solsticehedgehog',
+		url: 'https://mastodon.social/tags/solsticehedgehog',
+		history: [
+			{ day: '1671580800', accounts: '54', uses: '61' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '0', uses: '0' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+			{ day: '1671148800', accounts: '0', uses: '0' },
+			{ day: '1671062400', accounts: '0', uses: '0' },
+		],
+	},
+	{
+		name: 'waterfallwednesday',
+		url: 'https://mastodon.social/tags/waterfallwednesday',
+		history: [
+			{ day: '1671580800', accounts: '109', uses: '116' },
+			{ day: '1671494400', accounts: '4', uses: '4' },
+			{ day: '1671408000', accounts: '3', uses: '3' },
+			{ day: '1671321600', accounts: '1', uses: '1' },
+			{ day: '1671235200', accounts: '2', uses: '2' },
+			{ day: '1671148800', accounts: '5', uses: '6' },
+			{ day: '1671062400', accounts: '77', uses: '82' },
+		],
+	},
+	{
+		name: 'bautzen',
+		url: 'https://mastodon.social/tags/bautzen',
+		history: [
+			{ day: '1671580800', accounts: '50', uses: '62' },
+			{ day: '1671494400', accounts: '1', uses: '2' },
+			{ day: '1671408000', accounts: '5', uses: '11' },
+			{ day: '1671321600', accounts: '1', uses: '1' },
+			{ day: '1671235200', accounts: '2', uses: '2' },
+			{ day: '1671148800', accounts: '7', uses: '8' },
+			{ day: '1671062400', accounts: '9', uses: '13' },
+		],
+	},
+	{
+		name: 'googlefonts',
+		url: 'https://mastodon.social/tags/googlefonts',
+		history: [
+			{ day: '1671580800', accounts: '47', uses: '47' },
+			{ day: '1671494400', accounts: '0', uses: '0' },
+			{ day: '1671408000', accounts: '1', uses: '1' },
+			{ day: '1671321600', accounts: '0', uses: '0' },
+			{ day: '1671235200', accounts: '0', uses: '0' },
+			{ day: '1671148800', accounts: '0', uses: '0' },
+			{ day: '1671062400', accounts: '1', uses: '1' },
+		],
+	},
+]
