+++ frontend/src/utils/getErrorHtml/error-raw.html
@@ -7,6 +7,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="Wildebeest Error Page">
     <title>Wildebeest Error</title>
+    <meta name="og:title" content="Wildebeest Error">
+    <meta name="og:description" content="Wildebeest Error Page">
+    <meta name="og:type" content="website">
     <style>
     body {
         margin: 0;
