+++ frontend/src/utils/getNotFoundHtml/not-found-raw.html
@@ -7,6 +7,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="Wildebeest Page Not Found">
     <title>Wildebeest Not Found</title>
+    <meta name="og:title" content="Wildebeest Not Found">
+    <meta name="og:description" content="Wildebeest Page Not Found">
+    <meta name="og:type" content="website">
     <style>
     body {
         margin: 0;
