+++ ui-e2e-tests/invidivual-toot.spec.ts
@@ -2,21 +2,37 @@ import { test, expect } from '@playwright/test'
 
 test('Navigation to and view of an individual toot', async ({ page }) => {
 	await page.goto('http://127.0.0.1:8788/explore')
-	await page.locator('article').filter({ hasText: 'Ken White' }).locator('i.fa-globe + span').click()
+	await page.locator('article').filter({ hasText: 'Ben, just Ben' }).locator('i.fa-globe + span').click()
+	await expect(page.getByRole('button', { name: 'Back' })).toBeVisible()
+	await expect(page.getByRole('link', { name: 'Avatar of Ben, just Ben' })).toBeVisible()
+	await expect(page.getByRole('link', { name: 'Ben, just Ben', exact: true })).toBeVisible()
+	await expect(page.locator('span', { hasText: 'A very simple update: all good!' })).toBeVisible()
+})
 
-	const backButtonLocator = page.getByRole('button', { name: 'Back' })
-	await expect(backButtonLocator).toBeVisible()
-
-	const avatarLocator = page.locator('img[alt="Avatar of Ken White"]')
-	await expect(avatarLocator).toBeVisible()
-
-	const userLinkLocator = page.locator('a[href="/@Popehat"]', { hasText: 'Ken White' })
-	await expect(userLinkLocator).toBeVisible()
-
-	const tootContentLocator = page.locator('p', {
-		hasText: 'Just recorded the first Serious Trouble episode of the new year, out tomorrow.',
-	})
-	await expect(tootContentLocator).toBeVisible()
+test('Navigation to and view of an individual toot with images', async ({ page }) => {
+	await page.goto('http://127.0.0.1:8788/explore')
+	await page
+		.locator('article')
+		.filter({ hasText: "I'm Rafael and I am a web designer!" })
+		.locator('i.fa-globe + span')
+		.click()
+	await expect(page.getByRole('button', { name: 'Back' })).toBeVisible()
+	await expect(page.getByRole('link', { name: 'Avatar of Raffa123$' })).toBeVisible()
+	await expect(page.getByRole('link', { name: 'Raffa123$', exact: true })).toBeVisible()
+	await expect(page.locator('p', { hasText: "I'm Rafael and I am a web designer!" })).toBeVisible()
+	expect(await page.getByTestId('media-gallery').getByRole('img').count()).toBe(4)
+	await expect(page.getByTestId('images-modal')).not.toBeVisible()
+	await page.getByTestId('media-gallery').getByRole('img').nth(2).click()
+	await expect(page.getByTestId('images-modal')).toBeVisible()
+	for (const n of [2, 1, 0, 3, 2, 1, 0, 3]) {
+		await expect(page.getByTestId('images-modal').getByRole('img')).toHaveAttribute(
+			'src',
+			`https://loremflickr.com/640/480/abstract?lock=${100 + n}`
+		)
+		await page.getByTestId('left-btn').click()
+	}
+	await page.getByTestId('close-btn').click()
+	await expect(page.getByTestId('images-modal')).not.toBeVisible()
 })
 
 test("Navigation to and view of a toot's replies", async ({ page }) => {
@@ -24,38 +40,38 @@ test("Navigation to and view of a toot's replies", async ({ page }) => {
 
 	await page
 		.locator('article')
-		.filter({ hasText: 'Bethany Black' })
+		.filter({ hasText: 'George' })
 		.filter({
-			hasText: 'We did it! *wipes tear from eye*',
+			hasText: 'We did it!',
 		})
 		.locator('i.fa-globe + span')
 		.click()
 
 	await page
 		.locator('article')
-		.filter({ hasText: 'Zach Weinersmith' })
+		.filter({ hasText: 'Zak Smith' })
 		.filter({
 			hasText: 'Yes we did!',
 		})
 		.locator('i.fa-globe + span')
 		.click()
 
-	await expect(page.getByRole('link', { name: 'Avatar of Zach Weinersmith' })).toBeVisible()
-	await expect(page.getByRole('link', { name: 'Zach Weinersmith', exact: true })).toBeVisible()
+	await expect(page.getByRole('link', { name: 'Avatar of Zak Smith' })).toBeVisible()
+	await expect(page.getByRole('link', { name: 'Zak Smith', exact: true })).toBeVisible()
 	await expect(page.getByText('Yes we did!')).toBeVisible()
 
 	await page.getByRole('button', { name: 'Back' }).click()
 
 	await page
 		.locator('article')
-		.filter({ hasText: 'nixCraft' })
+		.filter({ hasText: 'Penny' })
 		.filter({
 			hasText: 'Yes you guys did it!',
 		})
 		.locator('i.fa-globe + span')
 		.click()
 
-	await expect(page.getByRole('link', { name: 'Avatar of nixCraft' })).toBeVisible()
-	await expect(page.getByRole('link', { name: 'nixCraft 🐧', exact: true })).toBeVisible()
+	await expect(page.getByRole('link', { name: 'Avatar of Penny' })).toBeVisible()
+	await expect(page.getByRole('link', { name: 'Penny', exact: true })).toBeVisible()
 	await expect(page.getByText('Yes you guys did it!')).toBeVisible()
 })
